refactor: Movable 제거
