+++ src/main/java/racingcar/domain/Car.java
@@ -1,21 +1,26 @@
 package racingcar.domain;
 
 
-public class Car extends Movable {
+public class Car {
 
     public record Name(String value) {
+
         public Name {
             if (value.length() > 5) {
                 throw new IllegalArgumentException("이름은 5자 이하여야 합니다");
             }
         }
+
     }
 
 
     public final Name name;
 
+    private final Environment environment;
+    private final Position position = new Position();
+
     public Car(Environment environment, Name name) {
-        super(environment);
+        this.environment = environment;
         this.name = name;
     }
 
@@ -23,4 +28,17 @@ public class Car extends Movable {
     public String getName() {
         return name.value;
     }
+
+
+    public int getPosition() {
+        return position.get();
+    }
+
+    public void tryAdvance() {
+        Environment.MoveResult result = environment.tryAdvance();
+        if (result == Environment.MoveResult.MOVE) {
+            position.forward();
+        }
+    }
+
 }
